# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pd-prep-for-pgdp",
    "description": "Convert a folder/zip of scanned book images into a PGDP-ready submission package.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/projects/{project_id}/pages": {
      "get": {
        "tags": [
          "pages"
        ],
        "summary": "List Pages",
        "operationId": "list_pages",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "page_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PageType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Type"
            }
          },
          {
            "name": "has_splits",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Splits"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PageProcessingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "review_needed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Needed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPagesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/projects/{project_id}/pages/{idx0}": {
      "get": {
        "tags": [
          "pages"
        ],
        "summary": "Get Page",
        "operationId": "get_page",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "idx0",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "pages"
        ],
        "summary": "Update Page",
        "operationId": "update_page",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "idx0",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx0"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/projects/{project_id}/pages/{idx0}/text": {
      "patch": {
        "tags": [
          "pages"
        ],
        "summary": "Update Page Text",
        "operationId": "update_page_text",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "idx0",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx0"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePageTextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/projects/{project_id}/pages/{idx0}/text/{suffix}": {
      "get": {
        "tags": [
          "pages"
        ],
        "summary": "Get Page Text",
        "operationId": "get_page_text",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "idx0",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx0"
            }
          },
          {
            "name": "suffix",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Suffix"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPageTextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/projects/{project_id}/pages/{idx0}/words": {
      "delete": {
        "tags": [
          "pages"
        ],
        "summary": "Delete Page Words",
        "description": "Hard-delete OCR words from a page's `<root>.words.json` and\nrewrite `<root>.txt` from the survivors (roadmap \u00a79a).\n\nv1 is intentionally a hard rewrite \u2014 the soft-delete-flag\nalternative is recorded in the roadmap as deferred. Unknown\n`word_ids` are silently skipped so the call is idempotent across\nretries; the response's `deleted_count` reports how many ids\nactually applied.",
        "operationId": "delete_page_words",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "idx0",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx0"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteWordsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWordsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/projects/{project_id}/pages/{idx0}/stages": {
      "get": {
        "tags": [
          "pages"
        ],
        "summary": "List Page Stages",
        "description": "Return ordered per-page stage state for the 22-stage DAG.\n\nSpec: `docs/specs/pipeline-task-model.md` \u00a7\"API surface\" (\u00a7Per-page\nstage routes). Lazy-init contract (Q1-followup): if no rows exist\nyet for this page, materialise 22 ``not-run`` rows in one\ntransaction (`INSERT OR IGNORE`) and return them in topological\norder. Concurrent first-touch reads converge on exactly 22 rows.\n\nAuth follows the existing pattern \u2014 every project read is filtered\nby `user.user_id`. 404 (not 403) is returned on miss to avoid leaking\nproject existence to non-owners.",
        "operationId": "list_page_stages",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "idx0",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PageStageState"
                  },
                  "title": "Response List Page Stages"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageProcessingStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "complete",
          "error"
        ],
        "title": "PageProcessingStatus"
      },
...
      "PageType": {
        "type": "string",
        "enum": [
          "normal",
          "blank",
          "plate_b",
          "plate_p",
          "plate_r"
        ],
        "title": "PageType"
      },
...
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": "default"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "proof_page_count": {
            "type": "integer",
            "title": "Proof Page Count"
          },
          "config": {
            "$ref": "#/components/schemas/ProjectConfig"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          },
          "storage_prefix": {
            "type": "string",
            "title": "Storage Prefix"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name",
          "created_at",
          "updated_at",
          "status",
          "page_count",
          "proof_page_count",
          "config",
          "pipeline_state",
          "storage_prefix",
          "archived"
        ],
        "title": "Project"
      },
...
      "UpdatePageTextRequest": {
        "properties": {
          "split_suffix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Suffix"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "UpdatePageTextRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}